Json Reader moved to proper directory
