+++ CultureInGdansk/Common/Events.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Text;
 
-namespace JSONparser
+namespace Common
 {
     public class Events
     {
